requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==11.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
